+++ lib/pragma/operation/base.rb
@@ -21,6 +21,54 @@ module Pragma
             .downcase
             .to_sym
         end
+
+        def respond(skill, *args)
+          if args.size == 1 && args.first.is_a?(Hash)
+            options = args[0]
+            response = Pragma::Operation::Response.new(status: options.fetch(:status, 200))
+          elsif args.size >= 1 && args.size <= 2
+            response = args[0]
+            options = args[1] || {}
+
+            unless response.is_a?(Pragma::Operation::Response)
+              response = Pragma::Operation::Response.const_get(response).new
+            end
+          else
+            fail ArgumentError, "Expected 2..3 arguments, #{args.size + 1} given"
+          end
+
+          response.entity = options[:entity] if options.key?(:entity)
+          response.headers = response.headers.merge(options[:headers]) if options[:headers]
+
+          response.decorate_with(options[:decorator]) if options[:decorator]
+
+          skill['result.response'] = response
+        end
+
+        def error(skill, *args)
+          if args.size == 1 && args.first.is_a?(Hash)
+            options = args[0]
+
+            error = Pragma::Operation::Error.new(
+              error_type: options.fetch(:error_type),
+              error_message: options.fetch(:error_message),
+              meta: options.fetch(:meta, {})
+            )
+          elsif args.size >= 1 && args.size <= 2
+            error = args[0]
+            options = args[1] || {}
+          else
+            fail ArgumentError, "Expected 2..3 arguments, #{args.size + 1} given"
+          end
+
+          if defined?(Pragma::Decorator::Error) && !options.key?(:decorator)
+            options[:decorator] = Pragma::Decorator::Error
+          end
+
+          response = Pragma::Operation::Response.const_get(error.error_type).new(entity: error)
+
+          respond(skill, response, options)
+        end
       end
     end
   end